Working on Projects Panel
Challenging to resolve single appearance text for SEO details of the project added to the same column
of the project containing logo and project summary so that when collapses at mobile size will
responsively collapse in correct order. However at desktop size this project detail needed to be
added outside of the row to display correctly.

This was resolved by adding an order flex property of 0 at mobile and 1-9 at desktop.
Overall the projects panel at desktop is:
row [ logo-whfnp order-1, proj-whfnp order-4, dtls-whfnp order-7,
      logo-foof order-2, proj-foof order-5, dtls-foof order-8,
      logo-crooked order-3, proj-crooked order-6, dtls-crooked order-9 ]

and for mobile is all set to order-0

Used a media query to set all display:flex order property to order:0 at mobile for default collapse
Media query sits inside Sass mix-in to loop through and set class .order-#{$i} to property of order-$i